+++ src/components/CVCNumbers/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useCVCNumbers } from '../../hooks/useCVCNumbers';
 import CVCNumbersView from './CVCNumbersView';
 
 export interface CVCNumbersProps {
@@ -8,34 +8,17 @@ export interface CVCNumbersProps {
   setCvcError: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-const CVC_NUMBERS_LENGTH = 3;
-const ERROR_MESSAGE = '숫자만 입력 가능합니다.';
-
 const CVCNumbers = ({
   cvcNumbers,
   setCvcNumbers,
   onComplete,
   setCvcError,
 }: CVCNumbersProps) => {
-  const [errorMessage, setErrorMessage] = useState('');
-  const [error, setError] = useState(false);
-
-  useEffect(() => {
-    const hasError = error || cvcNumbers.length !== 3;
-    setCvcError(hasError);
-  }, [error, cvcNumbers]);
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value;
-    if (/^[0-9]*$/.test(value) && value.length <= CVC_NUMBERS_LENGTH) {
-      setErrorMessage('');
-      setError(false);
-      setCvcNumbers(value);
-    } else {
-      setErrorMessage(ERROR_MESSAGE);
-      setError(true);
-    }
-  };
+  const { errorMessage, error, handleInputChange } = useCVCNumbers({
+    cvcNumbers,
+    setCvcNumbers,
+    setCvcError,
+  });
 
   return (
     <CVCNumbersView
